Reorganize and centralize robot hardware
This is a big change, but I think this will help a great deal.
Everything on the robot needs some kind of common initialization
boilerplate that's not specific to use case, so may as well separate it
out from the whole subsystem-based chungus and put it in a central class
for things like characterization robots.